학기 정보 db에서 받아오는 로직 수정 tabClock에서 0이 먼저 렌더링되는 것을 방지하기 위해 App에서 받아서 tabClock에 프롭으로 넘기는 것으로 수정
